// src/types/FormData.ts
export interface FormDataType {
  name: string
  email: string
  mobile: string
  street: string
  city: string
  country: string
  linkedin: string
  github: string
  positionOne: string
  skills: string
  organization: string
  positionTwo: string
  duration: [Date | null, Date | null]
  descriptionOne: string
  title: string
  link: string
  descriptionTwo: string
  college: string
  year: [Date | null, Date | null]
  qualification: string
  descriptionThree: string
}